requirements.txt
boto3==1.12.12
botocore==1.15.12
dj-database-url==0.5.0
Django==2.2.2
django-storages==1.9.1
docutils==0.15.2
gunicorn==19.9.0
jmespath==0.9.5
Pillow==7.0.0
psycopg2==2.8.3
python-dateutil==2.8.1
python-dotenv==0.10.3
python-http-client==3.2.5
pytz==2019.1
s3transfer==0.3.3
sendgrid==6.1.2
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1